this further optimizes the performance of rotation gates.

```jl
@benchmark apply!(r, $(put(5, 2=>Rx(0.5)))) setup=(r=rand_state(5))
```

before:

```julia
BenchmarkTools.Trial: 
  memory estimate:  688 bytes
  allocs estimate:  17
  --------------
  minimum time:     424.719 ns (0.00% GC)
  median time:      441.246 ns (0.00% GC)
  mean time:        473.802 ns (5.99% GC)
  maximum time:     10.010 μs (95.16% GC)
  --------------
  samples:          10000
  evals/sample:     199
```

after:

```julia
BenchmarkTools.Trial: 
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     72.515 ns (0.00% GC)
  median time:      92.314 ns (0.00% GC)
  mean time:        94.292 ns (0.72% GC)
  maximum time:     1.508 μs (93.35% GC)
  --------------
  samples:          10000
  evals/sample:     954
```