@GiggleLiu 2020-09-26
The register size is too small. Do you have benchmark  for the larger size?