@Roger-luo 2020-09-26
20 qubits

Before:

```jl
BenchmarkTools.Trial: 
  memory estimate:  688 bytes
  allocs estimate:  17
  --------------
  minimum time:     1.470 ms (0.00% GC)
  median time:      1.501 ms (0.00% GC)
  mean time:        1.500 ms (0.00% GC)
  maximum time:     1.558 ms (0.00% GC)
  --------------
  samples:          279
  evals/sample:     1
```

after:

```jl
BenchmarkTools.Trial: 
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     1.379 ms (0.00% GC)
  median time:      1.405 ms (0.00% GC)
  mean time:        1.407 ms (0.00% GC)
  maximum time:     1.489 ms (0.00% GC)
  --------------
  samples:          282
  evals/sample:     1
```